+++ src/IO.ts
@@ -91,7 +91,6 @@
  *
  * @since 2.0.0
  */
-import { ChainRec1 } from './ChainRec'
 import { identity } from './function'
 import { Monad1 } from './Monad'
 import { MonadIO1 } from './MonadIO'
@@ -217,18 +216,11 @@ export const monadIO: Monad1<URI> = {
 /**
  * @since 2.0.0
  */
-export const io: Monad1<URI> & MonadIO1<URI> & ChainRec1<URI> = {
+export const io: Monad1<URI> & MonadIO1<URI> = {
   URI,
   map: map_,
   of,
   ap: ap_,
   chain: chain_,
-  fromIO: identity,
-  chainRec: (a, f) => () => {
-    let e = f(a)()
-    while (e._tag === 'Left') {
-      e = f(e.left)()
-    }
-    return e.right
-  }
+  fromIO: identity
 }